Testing array of different types (int, short, long)

+++ BossPuzzle/Utils/Parser.cs
@@ -9,7 +9,7 @@ public static class Parser
         return Int32.Parse(s, NumberStyles.Integer, CultureInfo.InvariantCulture);
     }
 
-    public static short ToInt16(string s)
+    public static long ToInt16(string s)
     {
         return Int16.Parse(s, NumberStyles.Integer, CultureInfo.InvariantCulture);
     }
